+++ find_a_supplier/views.py
@@ -446,3 +446,46 @@ class IndustryLandingPageContactCMSSuccessView(SpecificRefererRequiredMixin, Bas
     @property
     def expected_referer_url(self):
         return reverse('find-a-supplier:industry-contact', kwargs=self.kwargs)
+
+
+class UnsubscribeView(CountryDisplayMixin, GA360Mixin, FormView):
+    form_class = forms.UnsubscribeForm
+    template_name = 'find_a_supplier/unsubscribe.html'
+    success_template_name = 'find_a_supplier/unsubscribe-success.html'
+    failure_template_name = 'find_a_supplier/unsubscribe-failure.html'
+
+    def __init__(self):
+        super().__init__()
+
+        self.set_ga360_payload(
+            page_id='FindASupplierAnonymousUnsubscribe',
+            business_unit='FindASupplier',
+            site_section='Notifications',
+            site_subsection='AnonymousUnsubscribe'
+        )
+
+    def get_initial(self):
+        if self.request.method == 'GET' and self.request.GET.get('email'):
+            return {'email': self.request.GET['email']}
+        return {}
+
+    def get(self, *args, **kwargs):
+        if not self.request.GET.get('email'):
+            return redirect('find-a-supplier:trade-home')
+        return super().get(*args, **kwargs)
+
+    def form_valid(self, form):
+        response = api_client.notifications.anonymous_unsubscribe(form.cleaned_data['email'])
+        if response.ok:
+            return TemplateResponse(
+                self.request,
+                self.success_template_name,
+                context=self.get_context_data()
+            )
+        elif response.status_code == 400:
+            return TemplateResponse(
+                self.request,
+                self.failure_template_name,
+                context=self.get_context_data()
+            )
+        response.raise_for_status()